src/main/java/io/hhplus/tdd/point/repository/PointHistoryRepositoryImpl.java
package io.hhplus.tdd.point.repository;

import io.hhplus.tdd.database.PointHistoryTable;
import io.hhplus.tdd.point.aggregate.domain.PointHistoryDomain;
import io.hhplus.tdd.point.aggregate.entity.PointHistory;
import io.hhplus.tdd.point.aggregate.vo.TransactionType;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
@RequiredArgsConstructor
public class PointHistoryRepositoryImpl implements PointHistoryRepository {

    private final PointHistoryTable pointHistoryTable;

    @Override
    public PointHistoryDomain save(long userId, long amount, TransactionType type, long updateMillis) {
        return pointHistoryTable.insert(userId, amount, type, updateMillis).toDomain();
    }

    @Override
    public List<PointHistoryDomain> findAllByUserId(long userId) {
        return pointHistoryTable.selectAllByUserId(userId).stream().map(PointHistory::toDomain).toList();
    }
}
